backend/requirements.txt
fastapi
uvicorn[standard]
sse-starlette
google-genai
Pillow
python-dotenv
pydantic
